Remove links to Tipeee service
Tipeee was setup to please audience from [fr] willing to pay with €uro, and with a page description in French. But, it also added a layer of complexity ( two services? why Patreon *and* Tipeee ?) and Tipeee audience is not big.  Even with a featuring on their official account on Twitter (ref: https://twitter.com/Tipeee_Officiel/status/611934043911778307 ) nothing rized. Thanks to the 3 or 4 Tipeeer on it who trusted the system !
The page is still active, I'll remove it in august. This commit unlink official website to the Tipeee page/service. 
I was happy to try :-) Now I know I should keep focus on a single service.