[new key] "Alternatives: "
As in "You can use this as a replacement" / "this is alternative choice for similar option". I'm using it on frontpage to propose new support method in a cleaner way. 

Note: It was hard to find a dictionnaire for Breton; and impossible for Silesian/Ślōnskŏ gŏdka ( I copied the Polish word on it, assuming its better than nothing ). Other translation are coming from Google Translate engine with the full source sentence ( with trailing options) to give context to the translation bot. Took me hours, sorry if its not precise ; I prefer this than adding a new key totally untranslated...